requirements.txt
fire==0.5.0
python-dotenv==1.0.0
torch==1.11.0
numpy==1.22.3
tqdm==4.63.0